add alert dialog for delete btn

+++ src/components/DeleteButton.tsx
@@ -3,6 +3,17 @@ import React, { useTransition } from "react";
 import { Button } from "./ui/button";
 import { Trash } from "lucide-react";
 import { cn } from "@/lib/utils";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 
 const DeleteButton = ({
   deleteAction,
@@ -15,14 +26,40 @@ const DeleteButton = ({
 }) => {
   const [isPending, startTransition] = useTransition();
   return (
-    <Button
-      variant="destructive"
-      className={cn(className, "")}
-      onClick={() => startTransition(() => deleteAction(id))}
-      disabled={isPending}
-    >
-      <Trash />
-    </Button>
+    <AlertDialog>
+      <AlertDialogTrigger asChild>
+        <Button
+          variant="destructive"
+          className={cn(className, "")}
+          disabled={isPending}
+        >
+          <Trash />
+        </Button>
+      </AlertDialogTrigger>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Bist du dir sicher?</AlertDialogTitle>
+          <AlertDialogDescription>
+            Das Löschen von Veranstaltungen, Schichten und Aufgaben kann nicht
+            wieder Rückgängig gemacht werden.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>Abbrechen</AlertDialogCancel>
+          <AlertDialogAction
+            asChild
+            className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
+          >
+            <Button
+              disabled={isPending}
+              onClick={() => startTransition(() => deleteAction(id))}
+            >
+              Löschen
+            </Button>
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   );
 };
